asrock: add support for setting MAC addresses from FRU eeprom data
The apply-fru-macaddr.service runs a little script that reads six bytes
(raw binary) from an arbitrary offset within an arbitrary file (both
determined by a machine-specific config file) and then sets the MAC
address of a given network device (a third config parameter) to those
six bytes and saving it as a u-boot environment variable so it can be
applied on boot in the future (thus only the first boot of a given BMC
should need to actually do anything in this script; for all subsequent
boots it should just be a double-check).  The MAC address is also
appended to the hostname so it can contain a device-unique identifier
instead of just having the machine type and nothing else.

Signed-off-by: Zev Weiss <zweiss@equinix.com>
Change-Id: I67ae33bdb1c46f76049a7e66a140f38fb2601a06